chore: move color to commons
